refactor: 모임 라운드 타입 통일

+++ src/main/java/com/example/doktoribackend/meeting/service/MeetingService.java
@@ -114,7 +114,7 @@ public class MeetingService {
                     Book book = resolveBook(bookRequest);
                     LocalDateTime startAt = LocalDateTime.of(round.getDate(), startTime);
                     LocalDateTime endAt = startAt.plusMinutes(durationMinutes);
-                    return MeetingRound.create(meeting, book, (byte) roundNo, startAt, endAt);
+                    return MeetingRound.create(meeting, book, roundNo, startAt, endAt);
                 })
                 .toList();
         meetingRoundRepository.saveAll(rounds);
